Add 'unset' (deregister) option to the image command environment.
